issue 10. fixtures, test, supplier_list supplier_id problem fixed
