+++ README.md
@@ -7,6 +7,18 @@ pip install -r requirements.txt
 
 #Запуск
 
+При первичном запуске произвести миграции
+
+```
+./manage.py migrate
+```
+
+При первичном запуске заполнить модель начальными данными
+
+```
+./manage.py loaddata rsqm\rsqm_fixture.json
+```
+
 При первичном запуске следует создать суперпользователя
 
 ```
