+++ rsqm/rsqm_fixture.json
@@ -1,0 +1,1 @@
+[{"model": "contenttypes.contenttype", "pk": 1, "fields": {"app_label": "admin", "model": "logentry"}}, {"model": "contenttypes.contenttype", "pk": 2, "fields": {"app_label": "auth", "model": "permission"}}, {"model": "contenttypes.contenttype", "pk": 3, "fields": {"app_label": "auth", "model": "group"}}, {"model": "contenttypes.contenttype", "pk": 4, "fields": {"app_label": "auth", "model": "user"}}, {"model": "contenttypes.contenttype", "pk": 5, "fields": {"app_label": "contenttypes", "model": "contenttype"}}, {"model": "contenttypes.contenttype", "pk": 6, "fields": {"app_label": "sessions", "model": "session"}}, {"model": "contenttypes.contenttype", "pk": 7, "fields": {"app_label": "supplier", "model": "email"}}, {"model": "contenttypes.contenttype", "pk": 8, "fields": {"app_label": "supplier", "model": "supplier"}}, {"model": "contenttypes.contenttype", "pk": 9, "fields": {"app_label": "supplier", "model": "warehouse"}}, {"model": "contenttypes.contenttype", "pk": 10, "fields": {"app_label": "accordance", "model": "match"}}, {"model": "contenttypes.contenttype", "pk": 11, "fields": {"app_label": "accordance", "model": "product"}}, {"model": "contenttypes.contenttype", "pk": 12, "fields": {"app_label": "accordance", "model": "quantity"}}, {"model": "sessions.session", "pk": "h3vgluyastfkk3dibpk3gzh6dww1mzpz", "fields": {"session_data": "NzFlMzk2N2I2MTkzYTg5OTcyNTAzNjEyN2QzMGFkZDM0NTk1OTYxYzp7Il9hdXRoX3VzZXJfaWQiOiIxIiwiX2F1dGhfdXNlcl9iYWNrZW5kIjoiZGphbmdvLmNvbnRyaWIuYXV0aC5iYWNrZW5kcy5Nb2RlbEJhY2tlbmQiLCJfYXV0aF91c2VyX2hhc2giOiJhNTM4ODQwMzIyYjI5MzZhM2U3NWZlZTM0YTAxODE4OTQ2NDlhNzNkIn0=", "expire_date": "2017-05-14T09:40:44.004Z"}}, {"model": "supplier.supplier", "pk": 1, "fields": {"name": "supplier 0"}}, {"model": "supplier.supplier", "pk": 2, "fields": {"name": "supplier 1"}}, {"model": "supplier.supplier", "pk": 3, "fields": {"name": "supplier 2"}}, {"model": "supplier.warehouse", "pk": 1, "fields": {"supplier": 1, "name": "warehouse 0", "city": "city 0"}}, {"model": "supplier.warehouse", "pk": 2, "fields": {"supplier": 2, "name": "warehouse 1", "city": "city 1"}}, {"model": "supplier.warehouse", "pk": 3, "fields": {"supplier": 3, "name": "warehouse 2", "city": "city 2"}}, {"model": "supplier.email", "pk": 1, "fields": {"supplier": 1, "email": "s@s0.ru"}}, {"model": "supplier.email", "pk": 2, "fields": {"supplier": 2, "email": "s@s1.ru"}}, {"model": "supplier.email", "pk": 3, "fields": {"supplier": 3, "email": "s@s2.ru"}}, {"model": "accordance.product", "pk": 1, "fields": {"code": 1001}}, {"model": "accordance.product", "pk": 2, "fields": {"code": 2002}}, {"model": "accordance.product", "pk": 3, "fields": {"code": 3003}}, {"model": "auth.permission", "pk": 1, "fields": {"name": "Can add log entry", "content_type": 1, "codename": "add_logentry"}}, {"model": "auth.permission", "pk": 2, "fields": {"name": "Can change log entry", "content_type": 1, "codename": "change_logentry"}}, {"model": "auth.permission", "pk": 3, "fields": {"name": "Can delete log entry", "content_type": 1, "codename": "delete_logentry"}}, {"model": "auth.permission", "pk": 4, "fields": {"name": "Can add permission", "content_type": 2, "codename": "add_permission"}}, {"model": "auth.permission", "pk": 5, "fields": {"name": "Can change permission", "content_type": 2, "codename": "change_permission"}}, {"model": "auth.permission", "pk": 6, "fields": {"name": "Can delete permission", "content_type": 2, "codename": "delete_permission"}}, {"model": "auth.permission", "pk": 7, "fields": {"name": "Can add group", "content_type": 3, "codename": "add_group"}}, {"model": "auth.permission", "pk": 8, "fields": {"name": "Can change group", "content_type": 3, "codename": "change_group"}}, {"model": "auth.permission", "pk": 9, "fields": {"name": "Can delete group", "content_type": 3, "codename": "delete_group"}}, {"model": "auth.permission", "pk": 10, "fields": {"name": "Can add user", "content_type": 4, "codename": "add_user"}}, {"model": "auth.permission", "pk": 11, "fields": {"name": "Can change user", "content_type": 4, "codename": "change_user"}}, {"model": "auth.permission", "pk": 12, "fields": {"name": "Can delete user", "content_type": 4, "codename": "delete_user"}}, {"model": "auth.permission", "pk": 13, "fields": {"name": "Can add content type", "content_type": 5, "codename": "add_contenttype"}}, {"model": "auth.permission", "pk": 14, "fields": {"name": "Can change content type", "content_type": 5, "codename": "change_contenttype"}}, {"model": "auth.permission", "pk": 15, "fields": {"name": "Can delete content type", "content_type": 5, "codename": "delete_contenttype"}}, {"model": "auth.permission", "pk": 16, "fields": {"name": "Can add session", "content_type": 6, "codename": "add_session"}}, {"model": "auth.permission", "pk": 17, "fields": {"name": "Can change session", "content_type": 6, "codename": "change_session"}}, {"model": "auth.permission", "pk": 18, "fields": {"name": "Can delete session", "content_type": 6, "codename": "delete_session"}}, {"model": "auth.permission", "pk": 19, "fields": {"name": "Can add email", "content_type": 7, "codename": "add_email"}}, {"model": "auth.permission", "pk": 20, "fields": {"name": "Can change email", "content_type": 7, "codename": "change_email"}}, {"model": "auth.permission", "pk": 21, "fields": {"name": "Can delete email", "content_type": 7, "codename": "delete_email"}}, {"model": "auth.permission", "pk": 22, "fields": {"name": "Can add supplier", "content_type": 8, "codename": "add_supplier"}}, {"model": "auth.permission", "pk": 23, "fields": {"name": "Can change supplier", "content_type": 8, "codename": "change_supplier"}}, {"model": "auth.permission", "pk": 24, "fields": {"name": "Can delete supplier", "content_type": 8, "codename": "delete_supplier"}}, {"model": "auth.permission", "pk": 25, "fields": {"name": "Can add warehouse", "content_type": 9, "codename": "add_warehouse"}}, {"model": "auth.permission", "pk": 26, "fields": {"name": "Can change warehouse", "content_type": 9, "codename": "change_warehouse"}}, {"model": "auth.permission", "pk": 27, "fields": {"name": "Can delete warehouse", "content_type": 9, "codename": "delete_warehouse"}}, {"model": "auth.permission", "pk": 28, "fields": {"name": "Can add match", "content_type": 10, "codename": "add_match"}}, {"model": "auth.permission", "pk": 29, "fields": {"name": "Can change match", "content_type": 10, "codename": "change_match"}}, {"model": "auth.permission", "pk": 30, "fields": {"name": "Can delete match", "content_type": 10, "codename": "delete_match"}}, {"model": "auth.permission", "pk": 31, "fields": {"name": "Can add product", "content_type": 11, "codename": "add_product"}}, {"model": "auth.permission", "pk": 32, "fields": {"name": "Can change product", "content_type": 11, "codename": "change_product"}}, {"model": "auth.permission", "pk": 33, "fields": {"name": "Can delete product", "content_type": 11, "codename": "delete_product"}}, {"model": "auth.permission", "pk": 34, "fields": {"name": "Can add quantity", "content_type": 12, "codename": "add_quantity"}}, {"model": "auth.permission", "pk": 35, "fields": {"name": "Can change quantity", "content_type": 12, "codename": "change_quantity"}}, {"model": "auth.permission", "pk": 36, "fields": {"name": "Can delete quantity", "content_type": 12, "codename": "delete_quantity"}}, {"model": "auth.user", "pk": 1, "fields": {"password": "pbkdf2_sha256$36000$43xuo47lVo2k$uxIItNa0g1v881wM2P5bgNk4SCFhZ6qoXh0flsFgz64=", "last_login": "2017-04-30T09:40:43.976Z", "is_superuser": true, "username": "giger", "first_name": "", "last_name": "", "email": "a@a.ru", "is_staff": true, "is_active": true, "date_joined": "2017-04-29T13:19:20.317Z", "groups": [], "user_permissions": []}}]
